refactor: setup errors-core
chore: cleanup
