+++ packages/errors-core/.gitignore
@@ -1,0 +1,2 @@
+.docs/
+dist/
