+++ packages/errors-core/.prettierignore
@@ -1,0 +1,4 @@
+# Changelogs are autogenerated, so leave them alone
+CHANGELOG.md
+
+dist/
