+++ packages/errors-core/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @solana/errors-core
