+++ packages/errors-core/src/.npmignore
@@ -1,0 +1,4 @@
+__benchmarks__/
+__mocks__/
+__tests__/
+__typetests__/
