+++ packages/errors-core/src/__typetests__/coded-error-typetest.ts
