+++ packages/errors-core/tsconfig.declarations.json
@@ -1,0 +1,10 @@
+{
+    "compilerOptions": {
+        "declaration": true,
+        "declarationMap": true,
+        "emitDeclarationOnly": true,
+        "outDir": "./dist/types"
+    },
+    "extends": "./tsconfig.json",
+    "include": ["../build-scripts/build-time-constants.d.ts", "src/index.ts"]
+}
