+++ packages/errors-core/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "$schema": "https://json.schemastore.org/tsconfig",
+    "compilerOptions": {
+        "lib": ["DOM", "ES2015", "ES2022.Error"],
+        "resolveJsonModule": true
+    },
+    "display": "@solana/errors-core",
+    "extends": "../tsconfig/base.json",
+    "include": ["../build-scripts/build-time-constants.d.ts", "src"]
+}
